refactor(web): SettingsSection を settings feature 配下へ移動する
設定ドメイン固有の表示コンポーネントのため、汎用レイアウト（components/ui/layout）
ではなく features/settings の公開口から提供する。
