+++ application/apps/web/src/client/features/settings/index.ts
@@ -1,0 +1,2 @@
+// 設定機能の公開口
+export { default as SettingsSection } from './components/settings-section'
